refactor: split worker session handle modules
